+++ wojilu.Apps/Blog/Interface/ISysBlogService.cs
@@ -11,10 +11,10 @@ namespace wojilu.Apps.Blog.Interface {
     public interface ISysBlogService {
 
         List<BlogPost> GetSysHit( int count );
-        List<BlogPost> GetSysNew( int categoryId, int count );
+        List<BlogPost> GetSysNew(long categoryId, int count);
 
         DataPage<BlogPost> GetSysPage( int size );
-        DataPage<BlogPost> GetSysPageByCategory( int categoryId, int size );
+        DataPage<BlogPost> GetSysPageByCategory(long categoryId, int size);
         DataPage<BlogPost> GetSysPageBySearch( String condition );
         DataPage<BlogPost> GetSysPageTrash();
         List<BlogPost> GetSysReply( int count );
@@ -28,7 +28,7 @@ namespace wojilu.Apps.Blog.Interface {
         void DeleteTrue( string ids );
         void UnDelete( string ids );
 
-        List<BlogPost> GetByCategory( int categoryId, int count );
+        List<BlogPost> GetByCategory(long categoryId, int count);
     }
 
 }
